Properly handle connection refusal in reassembly
When the first packet from one of the enpoints is RST/ACK, do not
buffer it waiting to see a packet with SYN flag, but process it
right away. When a server refuses to connect, it will replies with
RST/ACK and sends nothing more.
See https://github.com/google/gopacket/issues/425