Prevent RST packet from starting new connection
If reassembly finds a packet for which there is no connection
being tracked, it creates a new connection if this packet has SYN
flag set or if it has non-empty payload. Add one more condition
for this check - if packet has RST flag set, don't allow it to
start new conn (even if it has SYN and/or some data in payload).