Make Stream.Accept return enum instead of bools
Refactoring: add new enum PacketDecision and make method Accept
of Stream interface return it instead of two bools. This enum
can currently be KeepDecision (which tells reassembly to keep
this packet adding it to session), DropDecision (drop the packet
without adding it) and EndSessionDecision (add this packet to
session and flush and remove it).