Allow Stream.Accept to signal session restart
Add new value of PacketDecision - StartNewSessionDecision.
Returning this from Accept tells reassembly to flush and close
current session without adding current packet to it and then
re-process current packet as if no session for it existed (which
should normally lead to creation of new session).
This can be useful when we did not see old existing session close
but then new session starts reusing the same port numbers.