Report whether session chunk is given during flush
Stream.ReassembledSG now has an additional argument - 'flushing'.
Reassembly will pass 'true' in it if chunk is given by
'skipFlush' function which means it was not assembled by a normal
TCP reassembly process, but was flushed from buffer because
session is closing before a proper place was found for it.
This most likely means some session data was lost and session is
therefore partial.